test: added saving logic data to csv files and to chart.
The data created about the subset of avidians that do the choosen
logic functions. The data is now saved the csv file and that file
is now part of the populated dish freezer item and saved in the
workspace. When the populated dish is loaded in the grid the logic
funcitons are selected and the data shows up on the chart.

This data does NOT show up on the analysis page. I might add that later.