test: corrected mispelled name of menu item. mnCnPopRun
I had been looking at changing the control menu so that Population
Run/Pause would work on the population page and Organism Run/Pause
rather than the currrent system where there is a separate Run and Pause
command and it just works on the population page. I have not been able
to get the dijit/menuitem to let me change the text of the menu without
messing up the formating of the menu.